requirement.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-SocketIO
Twilio
bcrypt
python-dotenv